feat(mwpw-125698): importing lana instead of rewriting lana
