+++ react/src/js/app.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import ReactDOM, { render } from 'react-dom';
 import { DOMRegistry } from 'react-dom-components';
 import { parseToPrimitive } from './components/Consonant/Helpers/general';
-import { loadLana } from './components/Consonant/Helpers/hooks';
+import { loadLana } from './components/Consonant/Helpers/lana';
 import Container from './components/Consonant/Container/Container';
 import consonantPageRDC from './components/Consonant/Page/ConsonantPageDOM';
 
@@ -12,6 +12,8 @@ const domRegistry = new DOMRegistry(React, render);
 domRegistry.register({
     consonantPageRDC,
 });
+
+loadLana();
 
 const initReact = (element) => {
     domRegistry.init(element);
@@ -44,8 +46,6 @@ try {
 } catch (e) {
     /* eslint-disable no-empty */
 }
-
-loadLana({clientId: 'chimera' });
 
 export class ConsonantCardCollecton {
     constructor(config, element) {
