+++ react/src/js/components/Consonant/Helpers/lana.js
@@ -1,0 +1,25 @@
+export const loadLana = (options = {}) => {
+    if (window.lana) return;
+
+    const lanaError = (e) => {
+        if (window.lana) {
+            window.lana.log(e.reason || e.error || e.message, { errorType: 'i' });
+        }
+    };
+
+    window.lana = {
+        log: async (...args) => {
+            window.removeEventListener('error', lanaError);
+            window.removeEventListener('unhandledrejection', lanaError);
+            // eslint-disable-next-line import/no-unresolved, import/extensions
+            await fetch('https://milo.adobe.com/libs/utils/lana.js');
+            return window.lana.log(...args);
+        },
+        debug: false,
+        options,
+    };
+    window.addEventListener('error', lanaError);
+    window.addEventListener('unhandledrejection', lanaError);
+};
+
+export const logLana = ({ message, errorMessage, sampleRate } = {}) => `${message} | ${errorMessage} | ${sampleRate}`;
